# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/users": {
      "get": {
        "operationId": "UsersController_findAll",
        "summary": "List all users",
        "description": "Returns every user in the in-memory store. Accessible by any valid role.",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "RBAC role header. Accepted values: admin | community_manager | organizer | moderator | user. GET endpoints require any valid role. POST / PATCH / DELETE require admin.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "admin",
                "community_manager",
                "organizer",
                "moderator",
                "user"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of user records",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserDto"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Missing or invalid x-role header"
          }
        },
        "tags": [
          "users"
        ]
      },
      "post": {
        "operationId": "UsersController_create",
        "summary": "Create a new user",
        "description": "Creates a user record. Requires x-role: admin.",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "RBAC role header. Accepted values: admin | community_manager | organizer | moderator | user. GET endpoints require any valid role. POST / PATCH / DELETE require admin.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "admin",
                "community_manager",
                "organizer",
                "moderator",
                "user"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "User creation payload",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The newly created user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "403": {
            "description": "Only admin can create users"
          }
        },
        "tags": [
          "users"
        ]
      }
    },
    "/api/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "summary": "Get a single user by ID",
        "description": "Fetches one user record by their numeric ID.",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "RBAC role header. Accepted values: admin | community_manager | organizer | moderator | user. GET endpoints require any valid role. POST / PATCH / DELETE require admin.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "admin",
                "community_manager",
                "organizer",
                "moderator",
                "user"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Numeric user ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The matching user record",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "403": {
            "description": "Missing or invalid x-role header"
          },
          "404": {
            "description": "User not found"
          }
        },
        "tags": [
          "users"
        ]
      },
      "patch": {
        "operationId": "UsersController_update",
        "summary": "Update an existing user",
        "description": "Partially updates a user (all fields optional). Any authenticated role may call this to update their own profile (there is no per-request user identity beyond the x-role header in this project, so \"self\" is not server-enforced — the same trust model already used by communities/events update).",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "RBAC role header. Accepted values: admin | community_manager | organizer | moderator | user. GET endpoints require any valid role. POST / PATCH / DELETE require admin.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "admin",
                "community_manager",
                "organizer",
                "moderator",
                "user"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Numeric user ID to update",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Fields to update (all optional)",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated user record",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "403": {
            "description": "Missing or invalid x-role header"
          },
          "404": {
            "description": "User not found"
          }
        },
        "tags": [
          "users"
        ]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "summary": "Delete a user",
        "description": "Permanently deletes a user and cascades to their memberships, posts, and reports. Requires x-role: admin.",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "RBAC role header. Accepted values: admin | community_manager | organizer | moderator | user. GET endpoints require any valid role. POST / PATCH / DELETE require admin.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "admin",
                "community_manager",
                "organizer",
                "moderator",
                "user"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Numeric user ID to delete",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deletion confirmation",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "User 4 deleted"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Only admin can delete users"
          },
          "404": {
            "description": "User not found"
          }
        },
        "tags": [
          "users"
        ]
      }
    },
...
  },
  "info": {
    "title": "Gameunity Backend API",
    "description": "In-memory NestJS backend for frontend integration and academic evaluation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Dev Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "x-role": {
        "type": "apiKey",
        "in": "header",
        "name": "x-role",
        "description": "RBAC role header: admin | community_manager | moderator | user"
      }
    },
    "schemas": {
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "username": {
            "type": "string",
            "example": "player01"
          },
          "email": {
            "type": "string",
            "example": "player@example.com"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "owner",
              "community_manager",
              "organizer",
              "moderator",
              "user",
              "gamer"
            ],
            "example": "user"
          },
          "bio": {
            "type": "string",
            "example": "Casual and competitive gamer"
          }
        },
        "required": [
          "id",
          "username",
          "email",
          "role"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "minLength": 3,
            "maxLength": 30,
            "example": "newplayer"
          },
          "email": {
            "type": "string",
            "example": "newplayer@example.com"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "owner",
              "community_manager",
              "organizer",
              "moderator",
              "user",
              "gamer"
            ],
            "example": "user"
          },
          "bio": {
            "type": "string",
            "minLength": 5,
            "maxLength": 160,
            "example": "Loves MMORPG and FPS."
          },
          "firstName": {
            "type": "string",
            "example": "Alex"
          },
          "lastName": {
            "type": "string",
            "description": "May be empty for single-word full names",
            "example": "Morgan"
          },
          "phone": {
            "type": "string",
            "example": "+91 98765 43210"
          },
          "avatar": {
            "type": "string",
            "example": "/uploads/1234-avatar.png"
          }
        },
        "required": [
          "username",
          "email",
          "role"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "minLength": 3,
            "maxLength": 30,
            "example": "newplayer"
          },
          "email": {
            "type": "string",
            "example": "newplayer@example.com"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "owner",
              "community_manager",
              "organizer",
              "moderator",
              "user",
              "gamer"
            ],
            "example": "user"
          },
          "bio": {
            "type": "string",
            "minLength": 5,
            "maxLength": 160,
            "example": "Loves MMORPG and FPS."
          },
          "firstName": {
            "type": "string",
            "example": "Alex"
          },
          "lastName": {
            "type": "string",
            "description": "May be empty for single-word full names",
            "example": "Morgan"
          },
          "phone": {
            "type": "string",
            "example": "+91 98765 43210"
          },
          "avatar": {
            "type": "string",
            "example": "/uploads/1234-avatar.png"
          }
        }
      },
...
    }
  }
}